requirements.txt
beautifulsoup4==4.13.4
Django==5.2.5
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
google-api-python-client==2.176.0
google-auth-oauthlib==0.4.2
protobuf==6.31.1
requests==2.32.4
tldextract==5.3.0
gunicorn==20.1.0
mysqlclient==2.2.7
whitenoise==5.3.0
celery==5.5.3
redis==6.2.0
python-decouple==3.8
